+++ packages/ricos-content/web/lib/htmlConverters.ts
@@ -1,1 +1,0 @@
-export { fromHtml, toHtml } from '../src/converters/html';