+++ packages/ricos-content/web/src/converters/tiptap/fromTiptap/__tests__/fromTiptap.spec.ts
@@ -4,8 +4,8 @@ import migrationContentTiptap from '../../toTiptap/__tests__/migrationContentTip
 import draftMigrationContent from '../../../../../../../../e2e/tests/fixtures/migration-content.json';
 import { toTiptap, fromTiptap, tiptapToDraft, tiptapNodeDataToDraft } from '../..';
 import { getImageBlockData, getImageNode } from '../../utils';
-import { compare } from '../../../..';
-import { JSONContent } from '@tiptap-es5/core';
+import { compare } from '../../../../comparision/compare';
+import { JSONContent } from '@tiptap/core';
 
 describe('convert from Tiptap', () => {
   describe('to RichContent', () => {